update: extract waveform rendering
